refactor(t720): extract hashCode/rotatingPrompt helpers to utils/rotatingPrompt.ts
Moves inline pure utility functions to follow project convention of placing
reusable helpers in frontend/src/utils/ (per architecture-reviewer PASS WITH NOTES).
