+++ frontend/src/components/student/StudentOverviewTab.tsx
@@ -9,6 +9,7 @@ import { CefrBadge } from '@/components/dashboard/CefrBadge'
 import { getObjectiveUrgency, getDaysRemaining, formatDaysRemaining } from '@/lib/objectiveUrgency'
 import { SectionHeader } from './SectionHeader'
 import { formatMonthYear } from '@/utils/formatDate'
+import { rotatingPrompt } from '@/utils/rotatingPrompt'
 
 interface Props {
   student: Student
@@ -23,19 +24,6 @@ interface Props {
 // ---------------------------------------------------------------------------
 // Helpers
 // ---------------------------------------------------------------------------
-
-function hashCode(s: string): number {
-  let h = 0
-  for (let i = 0; i < s.length; i++) {
-    h = (Math.imul(31, h) + s.charCodeAt(i)) | 0
-  }
-  return h
-}
-
-function rotatingPrompt(studentId: string, prompts: string[]): string {
-  const today = new Date().toISOString().slice(0, 10)
-  return prompts[Math.abs(hashCode(studentId + today)) % prompts.length]
-}
 
 function getDisplayTitle(session: SessionLog): string {
   if (session.title && session.title !== 'Session') return session.title
